+++ src/progress.py
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+
 import subprocess
 import threading
 import os
